refactor(frontend): data validation helpers (#496)
`normaliseHeaders` is now mostly used for general validation of uploaded data, so this renames it to `dataValidation`.
